squeeze func impl and refactoring
