--- dpnp/dpnp_iface_arraymanipulation.py
@@ -1,79 +1,0 @@
-# cython: language_level=3
-# distutils: language = c++
-# -*- coding: utf-8 -*-
-# *****************************************************************************
-# Copyright (c) 2016-2020, Intel Corporation
-# All rights reserved.
-#
-# Redistribution and use in source and binary forms, with or without
-# modification, are permitted provided that the following conditions are met:
-# - Redistributions of source code must retain the above copyright notice,
-#   this list of conditions and the following disclaimer.
-# - Redistributions in binary form must reproduce the above copyright notice,
-#   this list of conditions and the following disclaimer in the documentation
-#   and/or other materials provided with the distribution.
-#
-# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-# AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-# IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
-# ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
-# LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
-# CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
-# SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
-# INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
-# CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
-# ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF
-# THE POSSIBILITY OF SUCH DAMAGE.
-# *****************************************************************************
-
-"""
-Interface of the array manipulation function of the dpnp
-
-Notes
------
-This module is a face or public interface file for the library
-it contains:
- - Interface functions
- - documentation for the functions
- - The functions parameters check
-
-"""
-
-
-import numpy
-import dpnp
-
-from dpnp.backend import *
-from dpnp.dparray import dparray
-from dpnp.dpnp_utils import *
-from dpnp.dpnp_iface_arraycreation import array
-
-__all__ = [
-    "asfarray",
-]
-
-
-def asfarray(a, dtype=numpy.float64):
-    """
-    Return an array converted to a float type.
-
-    For full documentation refer to :obj:`numpy.asfarray`.
-
-    Notes
-    -----
-    This function works exactly the same as :obj:`dpnp.array`.
-
-    """
-
-    if not use_origin_backend(a):
-        # behavior of original function: int types replaced with float64
-        if numpy.issubdtype(dtype, numpy.integer):
-            dtype = numpy.float64
-
-        # if type is the same then same object should be returned
-        if isinstance(a, dpnp.ndarray) and a.dtype == dtype:
-            return a
-
-        return array(a, dtype=dtype)
-
-    return call_origin(numpy.asfarray, a, dtype)
